% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryPoissonModel.R
\name{BinaryPoissonModelLogLik}
\alias{BinaryPoissonModelLogLik}
\title{Calculate the likelihood of a binary trait Poisson substituion model, 
given a tree and binary trait data at the tips.}
\usage{
BinaryPoissonModelLogLik(x, tree, x0, q01, q10, ord = reorder(tree, order
  = "postorder", index.only = TRUE))
}
\arguments{
\item{x}{an integer vector of size N, where N is the number of tips in tree.}

\item{tree}{a phylo object}

\item{x0, q01, q10}{model parameters.}

\item{ord}{indices of the rows in tree$edge in pruning order. Defaults to 
\code{reorder(tree, order = "postorder", index.only = TRUE))}}
}
\value{
the likelihood value.
}
\description{
The binary trait Poisson substitution model assumes that a 
trait (character) with two possible states (0 and 1) changes it's state at 
random with constant rates of substitution from 0 to 1 (q01) and from
1 to 0 (q10). The model parameters are the root-state x0, and the two rate
parameters q01 and q10. The likelihood for given model parameters <x0, q01, q10>
is defined as the probability of observing a given vector of states at the 
tips of a fixed phylogenetic tree.
}
