% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryPoissonModel.R
\name{BinaryPoissonModelLogLikCpp}
\alias{BinaryPoissonModelLogLikCpp}
\title{Calculate the likelihood of a Poisson binary trait substituion model, 
given a tree and binary trait data at the tips.}
\usage{
BinaryPoissonModelLogLikCpp(x, tree, x0, q01, q10,
  cppObject = NewBinaryPoissonModelCppObject(x, tree),
  mode = getOption("SPLITT.postorder.mode", 0))
}
\arguments{
\item{x}{an integer vector of size N, where N is the number of tips in tree.}

\item{tree}{a phylo object}

\item{x0}{model parameters.}

\item{q01}{model parameters.}

\item{q10}{model parameters.}

\item{cppObject}{a previously created object returned by \code{\link{NewBinaryPoissonModelCppObject}}}

\item{mode}{an integer denoting the mode for traversing the tree, i.e. serial vs parallel.}
}
\value{
the likelihood value.
}
\description{
Calculate the likelihood of a Poisson binary trait substituion model, 
given a tree and binary trait data at the tips.
}
