% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BinaryPoissonUsingSPLITT__BinaryPoissonModel__AlgorithmType}
\alias{BinaryPoissonUsingSPLITT__BinaryPoissonModel__AlgorithmType}
\alias{Rcpp_BinaryPoissonUsingSPLITT__BinaryPoissonModel__AlgorithmType-class}
\title{\code{TraversalAlgorithm}-type used in \code{BinaryPoissonModel}}
\description{
\code{TraversalAlgorithm}-type used in \code{BinaryPoissonModel}
}
