% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BinaryPoissonUsingSPLITT__BinaryPoissonModel__TraversalAlgorithm}
\alias{BinaryPoissonUsingSPLITT__BinaryPoissonModel__TraversalAlgorithm}
\alias{Rcpp_BinaryPoissonUsingSPLITT__BinaryPoissonModel__TraversalAlgorithm-class}
\title{Base class for \code{BinaryPoissonUsingSPLITT::BinaryPoissonModel::AlgorithmType}}
\description{
Base class for \code{BinaryPoissonUsingSPLITT::BinaryPoissonModel::AlgorithmType}
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{algorithm.}
}}

