% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BinaryPoissonUsingSPLITT__TraversalTaskBinaryPoissonModel}
\alias{BinaryPoissonUsingSPLITT__TraversalTaskBinaryPoissonModel}
\alias{Rcpp_BinaryPoissonUsingSPLITT__TraversalTaskBinaryPoissonModel-class}
\title{Rcpp module for the \code{TraversalTaskBinaryPoissonModel}-class}
\description{
Rcpp module for the \code{TraversalTaskBinaryPoissonModel}-class
}
