% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryPoissonModel.R
\name{NewBinaryPoissonModelCppObject}
\alias{NewBinaryPoissonModelCppObject}
\title{Create an instance of the Rcpp module for a given tree and trait data}
\usage{
NewBinaryPoissonModelCppObject(x, tree)
}
\arguments{
\item{x}{an integer vector of size N, where N is the number of tips in tree.}

\item{tree}{a phylo object}
}
\value{
an object to be passed as argument of the \link{BinaryPoissonModelLogLikCpp} function.
}
\description{
Create an instance of the Rcpp module for a given tree and trait data
}
\seealso{
\link{BinaryPoissonModelLogLikCpp}
}
